Maintain one e2e test file with pytest markers
